Move the compare functions into their own file
This sets us up for a larger refactor.
